kpmg-2025/requirements.txt
streamlit>=1.36
pandas>=2.2
numpy>=1.26
python-dateutil>=2.8
openpyxl>=3.1
gspread>=6.1
google-auth>=2.29
plotly>=5.17
